feat(verdict): store and accept float scores (0–10)
- Change Verdict.score column to double precision (Alembic 011)
- Update VerdictCreate/VerdictResponse and ORM model to float
- Align MCP post_verdict and SDK types with API

Made-with: Cursor

+++ backend/app/schemas/platform.py
@@ -150,7 +150,7 @@ class PaperResponse(PaperBase):
 class VerdictCreate(BaseModel):
     paper_id: uuid.UUID
     content_markdown: str = Field(..., min_length=1, description="Written assessment in markdown")
-    score: int = Field(..., ge=0, le=10, description="Score from 0 (reject) to 10 (strong accept)")
+    score: float = Field(..., ge=0, le=10, description="Score from 0 (reject) to 10 (strong accept)")
 
 
 class VerdictResponse(BaseModel):
@@ -160,7 +160,7 @@ class VerdictResponse(BaseModel):
     author_type: str
     author_name: Optional[str] = None
     content_markdown: str
-    score: int
+    score: float
     upvotes: int = 0
     downvotes: int = 0
     net_score: int = 0
